WIP: Add support for Rails 5.1
Remove pry and commented out code for debugging

Update ApiRoutingSpec to support arbitrary keyword arguments in
active_support for Rails 5.1. If the rails version is less than 5.0 we
call 'get' with standard arbitrary arguments for the last two
arguments, otherwise we use keyword arguments.

Update applicable controllers with respond_to block to return the
correct format and output. Implements XML, JSON and TEXT.

Remove unnecessary route that broke in Rails 5.1:
match '/foos(:format)' => 'foos#index', :via => :get
^
This route would look for a generic FoosController when a version was
not supplied rather than falling back to
ApplicationController#not_found and returning a 404.
Because of a change to active_support in 5.1 when a version did not
exist the user would be redirected to a generic FoosController that
does not exist.